requirements.txt
openai>=2.0.0
python-dotenv>=0.9.9